Started the GUI
Using the libary PyQt 5
when you run the program put 123 into the password bar and press enter to retrieve some info.
Old configs may not work

Did as many comments as possible to explain whats going on.
The code is very dirty atm I'm sorry.

Didn't have time to connect the output of the gui to the actual chrome design, it shouldn't be that bad to connect it although I think we may need to move some functions from the main class into a new one so that the window.py can access them.
Loading saves work fine
Creating new saves throws an error I think because I did not format the paydetails array correctly.

Apart from that there are some minor things to do, I have included it in comments at the top of the windows.py

If you want to create a new save comment out window.startWindow and create a new one using the Command Line.